% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{snake_caser}
\alias{snake_caser}
\title{Snake caser}
\usage{
snake_caser(x)
}
\arguments{
\item{x}{chr A character string}
}
\value{
A character string in snake case
}
\description{
A function for an RStudio Add-in to convert a sentence-like character string to a single element using snake case (i.e. words all lower case and separated by underscores). Useful for making variable names from natural language expressions.
}
