% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{to_snake_case}
\alias{to_snake_case}
\title{To snake case}
\usage{
to_snake_case(x)
}
\arguments{
\item{x}{chr A character string}
}
\value{
A character string in snake case
}
\description{
A function to convert a sentence-like character string to a single element using snake case (i.e. punctuation removed, words to all lower case and separated by underscores). Useful for making variable names from natural language expressions. Inspired by Bob Rudis' docxtractr::mcga( (\url{(https://github.com/hrbrmstr/docxtractr/})
}
\examples{
to_snake_case("snakes are venomous, not poisonous!")

}
